requirements/requirements.txt
Sklearn
Flask
flask_restful
python-binance
pickle
numpy
pandas
json
